+++ profiles/elvui.lua
@@ -3030,7 +3030,6 @@ local function SetupTrenchyUI()
     E.db.TrenchyUI.nameplates.castbarInterruptReady.r = 0
     E.db.TrenchyUI.nameplates.castbarMarkerColor.b = 0.10588236153126
     E.db.TrenchyUI.nameplates.castbarMarkerColor.r = 0
-    E.db.TrenchyUI.nameplates.classColorTargetIndicator = true
     E.db.TrenchyUI.nameplates.disableFriendlyHighlight = true
     E.db.TrenchyUI.nameplates.hideFriendlyRealm = true
     E.db.TrenchyUI.nameplates.focusGlow.color.a = 1